moved module component to controller directory
